lightning: refactor awaiting_channel_ready to transition factory
